wipeout types
