+++ ACadSharp/Entities/ClipType.cs
@@ -1,0 +1,17 @@
+﻿namespace ACadSharp.Entities
+{
+	/// <summary>
+	/// Clipping boundary type
+	/// </summary>
+	public enum ClipType : short
+	{
+		/// <summary>
+		/// Rectangular
+		/// </summary>
+		Rectangular = 1,
+		/// <summary>
+		/// Polygonal
+		/// </summary>
+		Polygonal = 2
+	}
+}
