+++ ACadSharp/Entities/Wipeout.cs
@@ -5,46 +5,6 @@ using System.Collections.Generic;
 
 namespace ACadSharp.Entities
 {
-	[Flags]
-	public enum ImageDisplayFlags : short
-	{
-		/// <summary>
-		/// None
-		/// </summary>
-		None = 0,
-		/// <summary>
-		/// Show image
-		/// </summary>
-		ShowImage = 1,
-		/// <summary>
-		/// Show image when not aligned with screen
-		/// </summary>
-		ShowNotAlignedImage = 2,
-		/// <summary>
-		/// Use clipping boundary
-		/// </summary>
-		UseClippingBoundary = 8,
-		/// <summary>
-		/// Transparency is on
-		/// </summary>
-		TransparencyIsOn = 8
-	}
-
-	/// <summary>
-	/// Clipping boundary type
-	/// </summary>
-	public enum ClipType : short
-	{
-		/// <summary>
-		/// Rectangular
-		/// </summary>
-		Rectangular = 1,
-		/// <summary>
-		/// Polygonal
-		/// </summary>
-		Polygonal = 2
-	}
-
 	/// <summary>
 	/// Represents a <see cref="Wipeout"/> entity.
 	/// </summary>
